Corrected a couple casting problems. Issue #8776.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@5799 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/invoice.cpp
@@ -510,7 +510,7 @@ void invoice::sSave()
 	     "WHERE (invchead_id=:invchead_id);" );
 
   q.bindValue(":invchead_id",			_invcheadid);
-  q.bindValue(":invchead_invcnumber",		_invoiceNumber->text().toInt());
+  q.bindValue(":invchead_invcnumber",		_invoiceNumber->text());
   q.bindValue(":invchead_cust_id",		_cust->id());
   q.bindValue(":invchead_invcdate",		_invoiceDate->date());
   q.bindValue(":invchead_shipdate",		_shipDate->date());
